Make part_id enumeration private
